Remove redundancy and clean up HTML and CSS
Removed some redundancy in the stylesheet and added classes in HTML
elements to assist with this. Added a few new classes for the stylesheet
to call on to allow any changes to easily be made to any font or
background choices. Went through HTML and stylesheet to find any code
that was unnecessary and did not have an effect on the final product.